requirements.txt
jax[cpu]
flax
optax
numpy
pandas
scikit-learn
fastapi
uvicorn
prefect
streamlit
python-dotenv
openai>=1.3.0
tqdm
requests
